Consolidate GPU text helpers into the retained renderer
