+++ Cargo.toml
@@ -11,7 +11,5 @@ members = [
     "crates/noon-typst",
     "crates/noon-text-native",
     "crates/noon-text-raster",
-    "crates/noon-text-atlas",
-    "crates/noon-text-render-wgpu",
 ]
 resolver = "2"
