+++ crates/noon-render-wgpu/Cargo.toml
@@ -13,19 +13,13 @@ native = [
     "wgpu/metal",
     "wgpu/gles",
     "wgpu/vulkan",
-    "noon-text-atlas/native",
-    "noon-text-render-wgpu/native",
 ]
 web = [
     "wgpu/webgpu",
     "wgpu/webgl",
-    "noon-text-atlas/web",
-    "noon-text-render-wgpu/web",
 ]
 ci-noop = [
     "wgpu/noop",
-    "noon-text-atlas/ci-noop",
-    "noon-text-render-wgpu/ci-noop",
 ]
 
 [dependencies]
@@ -33,8 +27,7 @@ bytemuck = { version = "1", features = ["derive"] }
 noon-core = { path = "../noon-core" }
 noon-geometry = { path = "../noon-geometry" }
 noon-runtime = { path = "../noon-runtime" }
-noon-text-atlas = { path = "../noon-text-atlas", default-features = false }
-noon-text-render-wgpu = { path = "../noon-text-render-wgpu", default-features = false }
+noon-text-raster = { path = "../noon-text-raster" }
 swash = "=0.2.10"
 wgpu = { version = "30.0.1", default-features = false, features = ["std", "wgsl"] }
