+++ crates/noon-render-wgpu/tests/glyph_atlas_residency_churn.rs
@@ -3,7 +3,7 @@
 use std::sync::Arc;
 
 use noon_core::{FontResourceHandle, FontResourceId};
-use noon_text_atlas::{GlyphAtlasEntry, GlyphAtlasPlane, GpuGlyphAtlas};
+use noon_render_wgpu::text::atlas::{GlyphAtlasEntry, GlyphAtlasPlane, GpuGlyphAtlas};
 use noon_text_raster::{
     GlyphRaster, GlyphRasterFormat, GlyphRasterImage, GlyphRasterKey, GlyphRasterPlacement,
 };
