+++ crates/noon-web/Cargo.toml
@@ -14,7 +14,6 @@ default = ["renderer"]
 renderer-smoke = ["renderer"]
 renderer = [
     "dep:noon-render-wgpu",
-    "dep:noon-text-render-wgpu",
     "dep:wgpu",
 ]
 
@@ -30,7 +29,6 @@ serde_json = "1"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 noon-render-wgpu = { path = "../noon-render-wgpu", default-features = false, features = ["web"], optional = true }
-noon-text-render-wgpu = { path = "../noon-text-render-wgpu", default-features = false, features = ["web"], optional = true }
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4"
 web-sys = { version = "0.3", features = ["Event", "EventTarget", "HtmlCanvasElement", "OffscreenCanvas", "Performance", "Window"] }
