+++ crates/noon-web/src/retained_execution_canvas.rs
@@ -1,12 +1,12 @@
 #[cfg(target_arch = "wasm32")]
 mod wasm {
     use noon_core::Vec2;
+    use noon_render_wgpu::text::TextDeviceMetrics;
     use noon_render_wgpu::{
         Camera2D, GpuRenderer, PathMeshPreload, RetainedFramePreparer, RetainedTextGpuState,
         UploadWrite,
     };
     use noon_runtime::FrameChanges;
-    use noon_text_render_wgpu::TextDeviceMetrics;
     use wasm_bindgen::prelude::*;
     use web_sys::OffscreenCanvas;
