+++ scripts/pr-risk-classifier.mjs
@@ -28,8 +28,8 @@ const retainedExecutionRustPaths = new Set([
 const hostUpdaterDiagnosticPaths = new Set([
   ".github/workflows/pr-fast.yml",
   "crates/noon-render-wgpu/src/gpu/retained_text.rs",
-  "crates/noon-text-render-wgpu/src/gpu.rs",
-  "crates/noon-text-render-wgpu/src/preparation.rs",
+  "crates/noon-render-wgpu/src/text/gpu.rs",
+  "crates/noon-render-wgpu/src/text/preparation.rs",
   "crates/noon-web/src/renderer_observation.rs",
   "crates/noon-web/src/renderer_observation/retained.rs",
   "crates/noon-web/src/retained_execution_canvas.rs",
@@ -44,7 +44,6 @@ const hostUpdaterDiagnosticPaths = new Set([
 
 const rendererCriticalPrefixes = Object.freeze([
   "crates/noon-render-wgpu/",
-  "crates/noon-text-render-wgpu/",
   "crates/noon-web/",
 ]);
